Refactored shell to use node files for directories
