+++ ush/ush_config.h
@@ -41,6 +41,7 @@ extern "C" {
 #define USH_CONFIG_ENABLE_COMMAND_XXD 1
 
 #define USH_CONFIG_ENABLE_FEATURE_COMMANDS 1
+#define USH_CONFIG_ENABLE_I2C_COMMANDS 1
 #define USH_CONFIG_ENABLE_FEATURE_AUTOCOMPLETE 1
 #define USH_CONFIG_ENABLE_FEATURE_SHELL_STYLES 1
